Fix rasc abstraction bugs
