+++ homeassistant/components/automation/__init__.py
@@ -12,8 +12,8 @@ import voluptuous as vol
 
 from homeassistant.components import websocket_api
 from homeassistant.components.blueprint import CONF_USE_BLUEPRINT
-from homeassistant.components.rascalscheduler.entity import BaseRoutineEntity
-from homeassistant.components.rascalscheduler.rascalscheduler import create_routine
+from homeassistant.components.rasc.entity import BaseRoutineEntity
+from homeassistant.components.rasc.scheduler import create_routine
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     ATTR_MODE,
