+++ homeassistant/components/automation/manifest.json
@@ -3,7 +3,7 @@
   "name": "Automation",
   "after_dependencies": ["device_automation", "webhook"],
   "codeowners": ["@home-assistant/core"],
-  "dependencies": ["blueprint", "trace", "script", "rascalscheduler"],
+  "dependencies": ["blueprint", "trace", "script", "rasc"],
   "documentation": "https://www.home-assistant.io/integrations/automation",
   "integration_type": "system",
   "quality_scale": "internal"
