--- homeassistant/components/rascalscheduler/manifest.json
@@ -1,9 +1,0 @@
-{
-  "domain": "rascalscheduler",
-  "name": "Rascal Scheduler",
-  "codeowners": ["@home-assistant/core"],
-  "config_flow": false,
-  "documentation": "https://github.com/home-assistant/core/blob/dev/README.rst",
-  "iot_class": "local_polling",
-  "issue_tracker": "https://www.home-assistant.io/help/"
-}
