+++ homeassistant/components/virtual/cover.py
@@ -82,7 +82,7 @@ async def async_setup_entry(
         hass, entry.data[ATTR_GROUP_NAME], PLATFORM_DOMAIN
     ):
         entity = COVER_SCHEMA(entity)
-        if CONF_COORDINATED in entity:
+        if entity[CONF_COORDINATED]:
             entities.append(CoordinatedVirtualCover(entity, coordinator))
         else:
             entities.append(VirtualCover(entity))
