+++ homeassistant/components/virtual/fan.py
@@ -74,7 +74,7 @@ async def async_setup_entry(
         hass, entry.data[ATTR_GROUP_NAME], PLATFORM_DOMAIN
     ):
         entity = FAN_SCHEMA(entity)
-        if CONF_COORDINATED in entity:
+        if entity[CONF_COORDINATED]:
             entities.append(CoordinatedVirtualFan(entity, coordinator))
         else:
             entities.append(VirtualFan(entity))
