+++ homeassistant/components/virtual/light.py
@@ -123,7 +123,7 @@ async def async_setup_entry(
         hass, entry.data[ATTR_GROUP_NAME], PLATFORM_DOMAIN
     ):
         entity = LIGHT_SCHEMA(entity)
-        if CONF_COORDINATED in entity:
+        if entity[CONF_COORDINATED]:
             entities.append(CoordinatedVirtualLight(entity, coordinator))
         else:
             entities.append(VirtualLight(entity))
@@ -158,12 +158,12 @@ class VirtualLight(VirtualEntity, LightEntity):
 
         self._attr_is_on = config.get(CONF_INITIAL_VALUE).lower() == STATE_ON
 
-        if self._attr_supported_features & ColorMode.BRIGHTNESS:
+        if ColorMode.BRIGHTNESS in self._attr_supported_color_modes:
             self._attr_brightness = config.get(CONF_INITIAL_BRIGHTNESS)
-        if self._attr_supported_features & ColorMode.HS:
+        if ColorMode.HS in self._attr_supported_color_modes:
             self._attr_color_mode = ColorMode.HS
             self._attr_hs_color = config.get(CONF_INITIAL_COLOR)
-        if self._attr_supported_features & ColorMode.COLOR_TEMP:
+        if ColorMode.COLOR_TEMP in self._attr_supported_color_modes:
             self._attr_color_mode = ColorMode.COLOR_TEMP
             self._attr_color_temp = config.get(CONF_INITIAL_COLOR_TEMP)
         if self._attr_supported_features & LightEntityFeature.EFFECT:
@@ -176,16 +176,16 @@ class VirtualLight(VirtualEntity, LightEntity):
 
         self._attr_is_on = state.state.lower() == STATE_ON
 
-        if self._attr_supported_features & ColorMode.BRIGHTNESS:
+        if ColorMode.BRIGHTNESS in self._attr_supported_color_modes:
             self._attr_brightness = state.attributes.get(
                 ATTR_BRIGHTNESS, config.get(CONF_INITIAL_BRIGHTNESS)
             )
-        if self._attr_supported_features & ColorMode.HS:
+        if ColorMode.HS in self._attr_supported_color_modes:
             self._attr_color_mode = ColorMode.HS
             self._attr_hs_color = state.attributes.get(
                 ATTR_HS_COLOR, config.get(CONF_INITIAL_COLOR)
             )
-        if self._attr_supported_features & ColorMode.COLOR_TEMP:
+        if ColorMode.COLOR_TEMP in self._attr_supported_color_modes:
             self._attr_color_mode = ColorMode.COLOR_TEMP
             self._attr_color_temp = state.attributes.get(
                 ATTR_COLOR_TEMP, config.get(CONF_INITIAL_COLOR_TEMP)
