+++ homeassistant/components/virtual/lock.py
@@ -81,7 +81,7 @@ async def async_setup_entry(
         hass, entry.data[ATTR_GROUP_NAME], PLATFORM_DOMAIN
     ):
         entity = LOCK_SCHEMA(entity)
-        if CONF_COORDINATED in entity:
+        if entity[CONF_COORDINATED]:
             entities.append(CoordinatedVirtualLock(hass, entity, coordinator))
         else:
             entities.append(VirtualLock(hass, entity))
