+++ homeassistant/components/virtual/switch.py
@@ -63,7 +63,7 @@ async def async_setup_entry(
         hass, entry.data[ATTR_GROUP_NAME], PLATFORM_DOMAIN
     ):
         entity = SWITCH_SCHEMA(entity)
-        if CONF_COORDINATED in entity:
+        if entity[CONF_COORDINATED]:
             entities.append(CoordinatedVirtualSwitch(entity, coordinator))
         else:
             entities.append(VirtualSwitch(entity))
